requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
google-genai
langchain_google_genai
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
streamlit
streamlit-float
arxiv
ddgs
pinecone
youtube_transcript_api